requirements.txt
fastapi
uvicorn
ultralytics
pillow
requests
psycopg2-binary